requirements.txt
ntplib
pillow